Preliminary Thing collision.
I can improve it by seeing which side of the collision box we hit when sliding from there
I dont know if I am tired.

Oh,I fixed a bug in Aiwnios,so be sure to pull and rebuild. I use a "static" CDoomLinedef for collidng with things. This could be imporoved but I am lazy.
Nrootconauto.